LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/credentialless/iframe-coep-require-corp.tentative.https-expected.txt

CONSOLE MESSAGE: Refused to display 'https://localhost:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,none)&uuid=48c5cd5d-1a04-4b22-b367-348626e3e72f' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://127.0.0.1:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,none)&uuid=dc38ab59-60a5-4ca7-846e-07f347cbb88b' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://127.0.0.1:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,credentialless)&uuid=3bc9c936-138c-44a8-b89a-22d5348db081' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,require-corp)&uuid=9b4fc8a0-7c1b-4ced-9055-368e754dc376 because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,credentialless)&uuid=7f8abf5e-9254-4d23-b288-ed44441c26aa' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,none)|header(Cross-Origin-Resource-Policy,cross-origin)&uuid=9bda0fd6-29fb-4c8f-acf4-04b6db9ebf7f' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://127.0.0.1:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,none)|header(Cross-Origin-Resource-Policy,cross-origin)&uuid=14602c93-2124-4abb-9de9-5f98cb106608' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://127.0.0.1:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,credentialless)|header(Cross-Origin-Resource-Policy,cross-origin)&uuid=c90ebc7d-77ed-486d-8b4f-5af18771d6f2' in a frame because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Refused to display 'https://localhost:9443/html/cross-origin-embedder-policy/credentialless/resources/executor.html?pipe=|header(Cross-Origin-Embedder-Policy,credentialless)|header(Cross-Origin-Resource-Policy,cross-origin)&uuid=95a65629-9a80-4aa5-81fc-ae7e7d8c2d52' in a frame because of Cross-Origin-Embedder-Policy.

Harness Error (TIMEOUT), message = null

PASS COEP:require-corp embeds same-origin COEP:none
PASS COEP:require-corp embeds cross-origin COEP:none
TIMEOUT COEP:require-corp embeds same-origin COEP:credentialless Test timed out
PASS COEP:require-corp embeds cross-origin COEP:credentialless
PASS COEP:require-corp embeds same-origin COEP:require-corp
PASS COEP:require-corp embeds cross-origin COEP:require-corp
PASS COEP:require-corp embeds same-origin COEP:none, CORP:cross-origin
PASS COEP:require-corp embeds cross-origin COEP:none, CORP:cross-origin
TIMEOUT COEP:require-corp embeds same-origin COEP:credentialless, CORP:cross-origin Test timed out
TIMEOUT COEP:require-corp embeds cross-origin COEP:credentialless, CORP:cross-origin Test timed out
PASS COEP:require-corp embeds same-origin COEP:require-corp, CORP:cross-origin
PASS COEP:require-corp embeds cross-origin COEP:require-corp, CORP:cross-origin
